feat: moved cdp logic to separate export
