+++ typescript/packages/x402-hono/src/index.ts
@@ -28,7 +28,7 @@ import { useFacilitator } from "x402/verify";
  * @param globalConfig.facilitatorUrl - URL of the payment facilitator service
  * @param globalConfig.address - Address to receive payments
  * @param globalConfig.network - Network identifier (e.g. 'base-sepolia')
- * @param globalConfig.createAuthHeaders - Function to create creates for the payment facilitator service. If using Coinbase's facilitator, use the createCdpAuthHeader function.
+ * @param globalConfig.createAuthHeaders - Function to create creates for the payment facilitator service..
  *
  * @returns A function that creates a Hono middleware handler for a specific payment amount
  *
