+++ typescript/packages/x402/package.json
@@ -64,6 +64,11 @@
       "import": "./dist/esm/client/index.js",
       "require": "./dist/cjs/client/index.js"
     },
+    "./cdp": {
+      "types": "./dist/esm/cdp/index.d.ts",
+      "import": "./dist/esm/cdp/index.js",
+      "require": "./dist/cjs/cdp/index.js"
+    },
     "./verify": {
       "types": "./dist/esm/verify/index.d.ts",
       "import": "./dist/esm/verify/index.js",
